console: Move dropdown into own component
