+++ pkg/webui/components/dropdown/dropdown.styl
@@ -1,0 +1,54 @@
+// Copyright © 2020 The Things Network Foundation, The Things Industries B.V.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+ul.dropdown
+  border-normal()
+  z-index: $zi.nav
+  right: 0
+  position: absolute
+  background: white
+  padding: 0
+  min-width: 14rem
+
+  li.dropdown-item
+    display: block
+    margin-bottom: 0
+
+    button
+      reset-button()
+
+    & > a, & > button
+      one-liner()
+      color: $tc-deep-gray
+      text-decoration: none
+      padding: $cs.m 0
+      width: 100%
+
+      span:last-child
+        nudge('down')
+
+    .icon
+      margin-left: $cs.m
+
+    &:not(:last-child)
+      border-normal('bottom')
+
+    &:hover
+      background-color: $c-active-blue
+
+      a, button
+        color: white
+
+      .icon
+        color: white
